+++ WormGame/Pooling/PoolableEntity.cs
@@ -4,43 +4,26 @@ using Otter.Graphics;
 namespace WormGame.Pooling
 {
     /// @author Antti Harju
-    /// @version 10.08.2020
+    /// @version 12.08.2020
     /// <summary>
-    /// Class for poolable entities.
+    /// Base class for poolable entities.
     /// </summary>
-    public class PoolableEntity : Otter.Core.Entity, IPoolable
+    public class PoolableEntity : Entity, IPoolable
     {
         /// <summary>
-        /// Disables Otter2D collision.
+        /// Default constructor.
         /// </summary>
-        public PoolableEntity()
+        public PoolableEntity(int id)
         {
             Collidable = false;
+            Id = id;
         }
-
-
-        /// <summary>
-        /// Entity identifier. Unique within the same pool.
-        /// </summary>
-        public int Id { get; set; }
-
-
-        /// <summary>
-        /// Set entity color.
-        /// </summary>
-        public virtual Color Color { get { return Graphic.Color ?? null; } set { Graphic.Color = value; } }
-
-
-        /// <summary>
-        /// Sets entity visibility and various other Otter2d entity properties.
-        /// </summary>
-        public virtual bool Enabled { get { return Visible; } set { AutoUpdate = value; AutoRender = value; Visible = value; } }
 
 
         /// <summary>
         /// Add entity to the scene.
         /// </summary>
-        /// <param name="scene"></param>
+        /// <param name="scene">Scene</param>
         public virtual void Add(Scene scene)
         {
             scene.Add(this);
@@ -48,11 +31,32 @@ namespace WormGame.Pooling
 
 
         /// <summary>
+        /// Entity color.
+        /// </summary>
+        public virtual Color Color { get { return Graphic.Color ?? null; } set { Graphic.Color = value; } }
+
+
+        /// <summary>
         /// Disable entity.
         /// </summary>
-        public virtual void Disable()
+        /// <param name="recursive">Disable recursively. False only when disabling is done by pooler.</param>
+        public virtual void Disable(bool recursive = true)
         {
-            Enabled = false;
+            AutoRender = false;
+            AutoUpdate = false;
+            Visible = false;
         }
+
+
+        /// <summary>
+        /// Is entity active.
+        /// </summary>
+        public bool Active { get { return Visible; } set { if (value) { AutoRender = true; AutoUpdate = true; Visible = true; } else Disable(); } }
+
+
+        /// <summary>
+        /// Identifier. Unique within the same pool.
+        /// </summary>
+        public int Id { get; }
     }
 }